Refactor Meta
