+++ sdk-core/src/main/java/care/data4life/sdk/model/Record.kt
@@ -16,18 +16,18 @@
 package care.data4life.sdk.model
 
 import care.data4life.fhir.stu3.model.DomainResource
-import care.data4life.sdk.model.definitions.Fhir3Record
+import care.data4life.sdk.model.ModelContract.Fhir3Record
 import java.util.Objects.hash
 
 open class Record<T : DomainResource>(
-        fhirResource: T,
-        meta: Meta?,
-        annotations: List<String>? = null
+    fhirResource: T,
+    meta: Meta?,
+    annotations: List<String>? = null
 ) : Fhir3Record<T> {
     constructor(fhirResource: T, meta: Meta?) : this(
-            fhirResource,
-            meta,
-            null
+        fhirResource,
+        meta,
+        null
     )
 
     override val identifier: String
